software/server/auth/requirements.txt
marshmallow==3.8.0
requests==2.22.0
peewee==3.13.3
Flask==1.1.2
Flask-Cors==3.0.9
flask-restplus==0.13.0
Flask-Script==2.0.6
peewee_db_evolve==3.7.6
app==0.0.1
PyJWT==1.7.1
python-dotenv==0.15.0
Werkzeug==0.16.1
gunicorn==20.0.4
PyMySQL==0.10.1
peewee-migrations==0.3.19
